pyproject.toml
[project]
name = "icn-flight-alert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
